Vite: Change popup build to a dynamic HTML source
HTML elements were initially planned to be completely handled by JS, so the HTML file was considered static and just copied over.

General layout elements are easier to add directly to the HTML file, so it was made as the actual bundle entry so that it can watched and rebundled during development.