Converters:  Refactor base converter to a concrete class and merge implementation from currency converter.
Initially, the main converter acted as both an abstract super class for converters while also having some useful implementations, like matching units and numbers. The currency class had a lot of extra methods and features which could be applied to any converter.

So these methods and features were moved to the main converter class, making it a functional concrete class which can then be extended and overriden if necessary or initialized as is with units and rates.