Controllers: Fix bug in controller view factory parameters
The lack of a default view settings object caused an attempt to destructure a property to, rightfully, throw.

Fix by adding a default value. Also fix the type definition related to the offending destructuring, and refactor the destructuring all together.

Finally, add a safety check for the classes array parameter since when passing a single class, it's not uncommon to forget to wrap it in an array. And the class being an iterable string, gets silently spread and added as a group of characters.