Content Scripts/Background: Update string extraction and DOM range mapping.
Initially strings were extracted from the selection by finding the text nodes that intersected with the selection's range. This helped map the conversions to ranges by comparing their match indices to the text node value. But, relying on individual nodes meant that strings that spanned multiple nodes (formatted text, multiple styles, etc) could only be matched against the values of the individual nodes. This didn't feel intuitive to experience and needed to be rethought.

This new approach relies on ranges instead of nodes. First the selection range is split into its contiguous visible text portions.

The selection range itself can be used directly in a perfect world, but some pages have script and style elements that ranges will include when serialized. On Amazon , in an arbitrary spot in the page, I found a range that serialized to a 112kB string 🙈.

After splitting the regular conversion (now cleaned up and looking much nicer) happens normally and the conversion ranges are found using simple node traversal with a text node tree walker.