Render: Conversion: Defer bubble size calculations from render to reveal.
Calculating the bubble dimensions off of the hidden duplicate was running when the outer conversion component rendered. This needs a lot of layout calculations and reflows especially when multiple conversions are rendered causing a the page to hang for a noticeable duration.

Instead of calculating on render, calculate before the bubble reveal.