Converters: Update matching to include full plural label indices.
Matching caught plural labels but excluded the trailing 's' when returning match indices. This caused on page unit highlighting, which uses those indices, look off since it excluded part of the word aka. [US Dollar]s.

Initially this was because of a design choice to keep the unit part of the match unchanged if it was to be used later (by match filters, etc), but that's not the right way to go. Instead, the source label, if needed, should be included alongside the match by the context that invokes the matcher, not somehow returned within the match itself.

So update the matcher, as well as the invoker to include the label and any dependent functions to use the included label instead of the matched label.