Content Scripts: Update DOM selection stringification and position mapping algorithm
Currently, the 3rd and best iteration of on page string recognition and position mapping The key is using the Selection object's stringifier.

Previously, ranges were used to stringify and map the indices of those strings, but ranges stringified nodes by returning their actual `data` property. Text nodes sometimes (but often enough to notice) contained hidden whitesapces that weren't rendered on screen. Sometimes within the same line or even splitting a word. This caused many supposed matches to get skipped, making the extension feel buggy and off

In contrast, Selection.toString() returns the 'rendered text' [W3C - Selection API] which thankfully ignores all those random whitespaces and also has the built in benefit of ignoring hidden text nodes like style and script elements (something that needed to accounted for in the previous iteration). Perfect right? There was just one caveat, mapping the output string, which isn't accessible with any other way to location in the document, to highlight the matches.

The function below handles that task as elegantly as I could get it within a day and a bit by traversing the input selection with a mini selection using Selection.modify() which, thankfully, closely (but not perfectly) matched the characters in the output string with eac step. The slight differences that do exist, mainly between element boundaries, are handle by the function.