Render: Auto scale pre-expansion amount container, centering the conv.
Amounts are rendered with a set width to align conversion units in the expanded view. This width carried over to the pre-expansion view even though it could be larger than the space needed by the shown amount, causing a variance in the apparent left padding of the conversion, making it to look off-center.

To recenter the conversion, allow the amount container to scale to the amount by removing the width property and enforce the expanded width as a max-width.