Content Scripts: Defer actions until no mouse button's pressed.
Wait for a mouse selection is completed before processing and rendering.

This was an indirect response to a bug where rendering conversions while the mouse was still held down causes the whole page to get selected since conversions are appended at the end of the document body but absolutely positioned at the selection putting the end of the selection after it.

But it also makes sense to conceptually implement since it's a unecessary to start processing while the user is still in the act of selecting.